Formatting & Type correction

+++ EntainNextToGo/EntainNextToGo/Data/RaceStartDate.swift
@@ -9,5 +9,5 @@ import Foundation
 
 struct RaceStartDate: Codable {
     /// Epoch time.
-    let seconds: Int
+    let seconds: Double
 }
